@@ -1,4 +1,4 @@
-go 1.23.4
+go 1.24
 
 module github.com/cri-o/cri-o
 
